+++ DQMOffline/Trigger/plugins/HTDQM.h
@@ -1,7 +1,7 @@
 #ifndef DQMOffline_Trigger_HTDQM_H
 #define DQMOffline_Trigger_HTDQM_H
 
-#include "DQMOffline/Trigger/plugins/TriggerDQMBase.h"
+#include "DQMOffline/Trigger/interface/TriggerDQMBase.h"
 #include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
 #include "DQMServices/Core/interface/DQMStore.h"
 #include "DataFormats/JetReco/interface/PFJet.h"
